core_crypto_ffi::context
Struct CoreCryptoContext

pub struct CoreCryptoContext { /* private fields */ }

Implementations

impl CoreCryptoContext

pub async fn e2ei_new_enrollment( &self, client_id: String, display_name: String, handle: String, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_new_enrollment

pub async fn e2ei_new_activation_enrollment( &self, display_name: String, handle: String, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_new_activation_enrollment

pub async fn e2ei_new_rotate_enrollment( &self, display_name: Option<String>, handle: Option<String>, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_new_rotate_enrollment

pub async fn e2ei_register_acme_ca( &self, trust_anchor_pem: String, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_register_acme_ca

pub async fn e2ei_register_intermediate_ca( &self, cert_pem: String, ) -> Result<NewCrlDistributionPoints, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_register_intermediate_ca_pem

pub async fn e2ei_register_crl( &self, crl_dp: String, crl_der: Vec<u8>, ) -> Result<CrlRegistration, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_register_crl

pub async fn e2ei_mls_init_only( &self, enrollment: Arc<E2eiEnrollment>, certificate_chain: String, nb_key_package: Option<u32>, ) -> Result<NewCrlDistributionPoints, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_mls_init_only

pub async fn e2ei_rotate( &self, conversation_id: Vec<u8>, ) -> Result<(), CoreCryptoError>

See core_crypto::mls::conversation::ConversationGuard::e2ei_rotate

pub async fn save_x509_credential( &self, enrollment: Arc<E2eiEnrollment>, certificate_chain: String, ) -> Result<NewCrlDistributionPoints, CoreCryptoError>

See core_crypto::context::CentralContext::save_x509_credential

pub async fn delete_stale_key_packages( &self, ciphersuite: Ciphersuite, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::delete_stale_key_packages

pub async fn e2ei_enrollment_stash( &self, enrollment: Arc<E2eiEnrollment>, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_enrollment_stash

pub async fn e2ei_enrollment_stash_pop( &self, handle: Vec<u8>, ) -> Result<E2eiEnrollment, CoreCryptoError>

See core_crypto::context::CentralContext::e2ei_enrollment_stash_pop

pub async fn e2ei_conversation_state( &self, conversation_id: Vec<u8>, ) -> Result<E2eiConversationState, CoreCryptoError>

See [core_crypto::mls::conversation::conversation_guard::ConversationGuard::e2ei_conversation_state]

pub async fn e2ei_dump_pki_env( &self, ) -> Result<Option<E2eiDumpedPkiEnv>, CoreCryptoError>

pub async fn e2ei_is_pki_env_setup(&self) -> Result<bool, CoreCryptoError>

See [core_crypto::mls::Client::e2ei_is_pki_env_setup]

pub async fn e2ei_is_enabled( &self, ciphersuite: Ciphersuite, ) -> Result<bool, CoreCryptoError>

See [core_crypto::mls::Client::e2ei_is_enabled]

pub async fn get_device_identities( &self, conversation_id: Vec<u8>, device_ids: Vec<ClientId>, ) -> Result<Vec<WireIdentity>, CoreCryptoError>

See [core_crypto::mls::Client::get_device_identities]

pub async fn get_user_identities( &self, conversation_id: Vec<u8>, user_ids: Vec<String>, ) -> Result<HashMap<String, Vec<WireIdentity>>, CoreCryptoError>

See [core_crypto::mls::Client::get_user_identities]

pub async fn get_credential_in_use( &self, group_info: Vec<u8>, credential_type: MlsCredentialType, ) -> Result<E2eiConversationState, CoreCryptoError>

See [core_crypto::mls::Client::get_credential_in_use]

impl CoreCryptoContext

pub async fn proteus_init(&self) -> Result<(), CoreCryptoError>

See core_crypto::proteus::ProteusCentral::try_new

pub async fn proteus_session_from_prekey( &self, session_id: String, prekey: Vec<u8>, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::proteus_session_from_prekey

pub async fn proteus_session_from_message( &self, session_id: String, envelope: Vec<u8>, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_session_from_message

pub async fn proteus_session_save( &self, session_id: String, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::proteus_session_save Note: This isnโt usually needed as persisting sessions happens automatically when decrypting/encrypting messages and initializing Sessions

pub async fn proteus_session_delete( &self, session_id: String, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::proteus_session_delete

pub async fn proteus_session_exists( &self, session_id: String, ) -> Result<bool, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_session_exists

pub async fn proteus_decrypt( &self, session_id: String, ciphertext: Vec<u8>, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_decrypt

pub async fn proteus_encrypt( &self, session_id: String, plaintext: Vec<u8>, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_encrypt

pub async fn proteus_encrypt_batched( &self, sessions: Vec<String>, plaintext: Vec<u8>, ) -> Result<HashMap<String, Vec<u8>>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_encrypt_batched

pub async fn proteus_new_prekey( &self, prekey_id: u16, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_new_prekey

pub async fn proteus_new_prekey_auto( &self, ) -> Result<ProteusAutoPrekeyBundle, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_new_prekey_auto

pub async fn proteus_last_resort_prekey( &self, ) -> Result<Vec<u8>, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_last_resort_prekey

pub fn proteus_last_resort_prekey_id(&self) -> Result<u16, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_last_resort_prekey_id

pub async fn proteus_fingerprint(&self) -> Result<String, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_fingerprint

pub async fn proteus_fingerprint_local( &self, session_id: String, ) -> Result<String, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_fingerprint_local

pub async fn proteus_fingerprint_remote( &self, session_id: String, ) -> Result<String, CoreCryptoError>

See core_crypto::context::CentralContext::proteus_fingerprint_remote

pub fn proteus_fingerprint_prekeybundle( &self, prekey: Vec<u8>, ) -> Result<String, CoreCryptoError>

See core_crypto::proteus::ProteusCentral::fingerprint_prekeybundle NOTE: uniffi doesnโt support associated functions, so we have to have the self here

pub async fn proteus_cryptobox_migrate( &self, path: String, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::proteus_cryptobox_migrate

impl CoreCryptoContext

pub async fn set_data(&self, data: Vec<u8>) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::set_data.

pub async fn get_data(&self) -> Result<Option<Vec<u8>>, CoreCryptoError>

See core_crypto::context::CentralContext::get_data.

pub async fn mls_init( &self, client_id: ClientId, ciphersuites: Ciphersuites, nb_key_package: Option<u32>, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::mls_init

pub async fn mls_generate_keypairs( &self, ciphersuites: Ciphersuites, ) -> Result<Vec<ClientId>, CoreCryptoError>

See core_crypto::context::CentralContext::mls_generate_keypairs

pub async fn mls_init_with_client_id( &self, client_id: ClientId, tmp_client_ids: Vec<ClientId>, ciphersuites: Ciphersuites, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::mls_init_with_client_id

pub async fn client_public_key( &self, ciphersuite: Ciphersuite, credential_type: MlsCredentialType, ) -> Result<Vec<u8>, CoreCryptoError>

See [core_crypto::mls::Client::client_public_key]

pub async fn conversation_epoch( &self, conversation_id: Vec<u8>, ) -> Result<u64, CoreCryptoError>

See [core_crypto::mls::conversation::ConversationGuard::epoch]

pub async fn conversation_ciphersuite( &self, conversation_id: &ConversationId, ) -> Result<Ciphersuite, CoreCryptoError>

See [core_crypto::mls::conversation::ConversationGuard::ciphersuite]

pub async fn conversation_exists( &self, conversation_id: Vec<u8>, ) -> Result<bool, CoreCryptoError>

See [core_crypto::mls::Client::conversation_exists]

pub async fn get_client_ids( &self, conversation_id: Vec<u8>, ) -> Result<Vec<ClientId>, CoreCryptoError>

See [core_crypto::mls::conversation::ImmutableConversation::get_client_ids]

pub async fn export_secret_key( &self, conversation_id: Vec<u8>, key_length: u32, ) -> Result<Vec<u8>, CoreCryptoError>

See [core_crypto::mls::conversation::ImmutableConversation::export_secret_key]

pub async fn get_external_sender( &self, conversation_id: Vec<u8>, ) -> Result<Vec<u8>, CoreCryptoError>

See [core_crypto::mls::conversation::ImmutableConversation::get_external_sender]

pub async fn client_keypackages( &self, ciphersuite: Ciphersuite, credential_type: MlsCredentialType, amount_requested: u32, ) -> Result<Vec<Vec<u8>>, CoreCryptoError>

See core_crypto::context::CentralContext::get_or_create_client_keypackages

pub async fn client_valid_keypackages_count( &self, ciphersuite: Ciphersuite, credential_type: MlsCredentialType, ) -> Result<u64, CoreCryptoError>

See core_crypto::context::CentralContext::client_valid_key_packages_count

pub async fn delete_keypackages( &self, refs: Vec<Vec<u8>>, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::delete_keypackages

pub async fn create_conversation( &self, conversation_id: Vec<u8>, creator_credential_type: MlsCredentialType, config: ConversationConfiguration, ) -> Result<(), CoreCryptoError>

See core_crypto::context::CentralContext::new_conversation

pub async fn process_welcome_message( &self, welcome_message: Vec<u8>, custom_configuration: CustomConfiguration, ) -> Result<WelcomeBundle, CoreCryptoError>

See core_crypto::context::CentralContext::process_raw_welcome_message

pub async fn add_clients_to_conversation( &self, conversation_id: Vec<u8>, key_packages: Vec<Vec<u8>>, ) -> Result<NewCrlDistributionPoints, CoreCryptoError>

See [core_crypto::mls::conversation::conversation_guard::ConversationGuard::add_members]

pub async fn remove_clients_from_conversation( &self, conversation_id: Vec<u8>, clients: Vec<ClientId>, ) -> Result<(), CoreCryptoError>

See [core_crypto::context::CentralContext::remove_members_from_conversation]

pub async fn mark_conversation_as_child_of( &self, child_id: Vec<u8>, parent_id: Vec<u8>, ) -> Result<(), CoreCryptoError>

See core_crypto::mls::conversation::ConversationGuard::mark_as_child_of

pub async fn update_keying_material( &self, conversation_id: Vec<u8>, ) -> Result<(), CoreCryptoError>

See [core_crypto::context::CentralContext::update_keying_material]

pub async fn commit_pending_proposals( &self, conversation_id: Vec<u8>, ) -> Result<(), CoreCryptoError>

See [core_crypto::mls::conversation::conversation_guard::ConversationGuard::commit_pending_proposals]

pub async fn wipe_conversation( &self, conversation_id: Vec<u8>, ) -> Result<(), CoreCryptoError>

see core_crypto::context::CentralContext::wipe_conversation

pub async fn decrypt_message( &self, conversation_id: Vec<u8>, payload: Vec<u8>, ) -> Result<DecryptedMessage, CoreCryptoError>

See [core_crypto::mls::conversation::conversation_guard::ConversationGuard::decrypt_message]

pub async fn encrypt_message( &self, conversation_id: Vec<u8>, message: Vec<u8>, ) -> Result<Vec<u8>, CoreCryptoError>

See [core_crypto::mls::conversation::conversation_guard::ConversationGuard::encrypt_message]

pub async fn join_by_external_commit( &self, group_info: Vec<u8>, custom_configuration: CustomConfiguration, credential_type: MlsCredentialType, ) -> Result<WelcomeBundle, CoreCryptoError>

See core_crypto::context::CentralContext::join_by_external_commit