core_crypto_ffi

Function x509_credential_acquisition_new_from_credential_ref

pub async fn x509_credential_acquisition_new_from_credential_ref(
    pki_environment: Arc<PkiEnvironment>,
    config: X509CredentialAcquisitionConfiguration,
    credential_ref: &CredentialRef,
    core_crypto_database: Option<Arc<Database>>,
) -> CoreCryptoResult<X509CredentialAcquisition>

Create a new credential acquisition from an existing credential. This API is temporary until our system decouples client identities from a client’s public signature key. See https://wearezeta.atlassian.net/wiki/x/RABtrQ.

Provide core_crypto_database if you’re using distinct DB instances for PkiEnvironment and CoreCrypto. Otherwise, the PkiEnvironment’s DB will be used to load the full credential.