Struct core_crypto_ffi::CoreCrypto

pub struct CoreCrypto<'a> { /* private fields */ }

Implementations

impl CoreCrypto<'_>

pub fn new<'s>(
path: &'s str,
key: &'s str,
client_id: &'s ClientId,
entropy_seed: Option<Vec<u8>>
) -> CryptoResult<Self>

See [core_crypto::MlsCentral::try_new]

pub fn deferred_init<'s>(
path: &'s str,
key: &'s str,
entropy_seed: Option<Vec<u8>>
) -> CryptoResult<Self>

Similar to CoreCrypto::new but defers MLS initialization. It can be initialized later with [CoreCrypto::init_mls]. This should stay as long as proteus is supported. Then it should be removed.

pub fn mls_init(&self, client_id: &ClientId) -> CryptoResult<()>

See [core_crypto::MlsCentral::init_mls]

pub fn restore_from_disk(&self) -> CryptoResult<()>

See [core_crypto::MlsCentral::restore_from_disk]

pub fn close(self) -> CryptoResult<()>

See [core_crypto::MlsCentral::close]

pub fn wipe(self) -> CryptoResult<()>

See [core_crypto::MlsCentral::wipe]

pub fn client_public_key(&self) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::client_public_key]

pub fn client_keypackages(
&self,
amount_requested: u32
) -> CryptoResult<Vec<Vec<u8>>>

See [core_crypto::MlsCentral::client_keypackages]

pub fn client_valid_keypackages_count(&self) -> CryptoResult<u64>

See [core_crypto::MlsCentral::client_valid_keypackages_count]

pub fn create_conversation(
&self,
conversation_id: ConversationId,
config: ConversationConfiguration
) -> CryptoResult<()>

See [core_crypto::MlsCentral::new_conversation]

pub fn conversation_epoch(
&self,
conversation_id: ConversationId
) -> CryptoResult<u64>

See [core_crypto::MlsCentral::conversation_epoch]

pub fn process_welcome_message(
&self,
welcome_message: &[u8],
custom_configuration: CustomConfiguration
) -> CryptoResult<ConversationId>

See [core_crypto::MlsCentral::process_raw_welcome_message]

pub fn add_clients_to_conversation(
&self,
conversation_id: ConversationId,
clients: Vec<Invitee>
) -> CryptoResult<MemberAddedMessages>

See [core_crypto::MlsCentral::add_members_to_conversation]

pub fn remove_clients_from_conversation(
&self,
conversation_id: ConversationId,
clients: Vec<ClientId>
) -> CryptoResult<CommitBundle>

See [core_crypto::MlsCentral::remove_members_from_conversation]

pub fn update_keying_material(
&self,
conversation_id: ConversationId
) -> CryptoResult<CommitBundle>

See [core_crypto::MlsCentral::update_keying_material]

pub fn commit_pending_proposals(
&self,
conversation_id: ConversationId
) -> CryptoResult<Option<CommitBundle>>

See [core_crypto::MlsCentral::commit_pending_proposals]

pub fn wipe_conversation(
&self,
conversation_id: ConversationId
) -> CryptoResult<()>

see [core_crypto::MlsCentral::wipe_conversation]

pub fn decrypt_message(
&self,
conversation_id: ConversationId,
payload: &[u8]
) -> CryptoResult<DecryptedMessage>

See [core_crypto::MlsCentral::decrypt_message]

pub fn encrypt_message(
&self,
conversation_id: ConversationId,
message: &[u8]
) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::encrypt_message]

pub fn conversation_exists(&self, conversation_id: ConversationId) -> bool

See [core_crypto::MlsCentral::conversation_exists]

pub fn new_add_proposal(
&self,
conversation_id: ConversationId,
keypackage: Vec<u8>
) -> CryptoResult<ProposalBundle>

See [core_crypto::MlsCentral::new_proposal]

pub fn new_update_proposal(
&self,
conversation_id: ConversationId
) -> CryptoResult<ProposalBundle>

See [core_crypto::MlsCentral::new_proposal]

pub fn new_remove_proposal(
&self,
conversation_id: ConversationId,
client_id: ClientId
) -> CryptoResult<ProposalBundle>

See [core_crypto::MlsCentral::new_proposal]

pub fn new_external_add_proposal(
&self,
conversation_id: ConversationId,
epoch: u64
) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::new_external_add_proposal]

pub fn new_external_remove_proposal(
&self,
conversation_id: ConversationId,
epoch: u64,
keypackage_ref: Vec<u8>
) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::new_external_remove_proposal]

pub fn export_group_state(
&self,
conversation_id: ConversationId
) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::export_public_group_state]

pub fn join_by_external_commit(
&self,
public_group_state: Vec<u8>,
custom_configuration: CustomConfiguration
) -> CryptoResult<ConversationInitBundle>

See [core_crypto::MlsCentral::join_by_external_commit]

pub fn merge_pending_group_from_external_commit(
&self,
conversation_id: ConversationId
) -> CryptoResult<()>

See [core_crypto::MlsCentral::merge_pending_group_from_external_commit]

pub fn clear_pending_group_from_external_commit(
&self,
conversation_id: ConversationId
) -> CryptoResult<()>

See [core_crypto::MlsCentral::clear_pending_group_from_external_commit]

pub fn random_bytes(&self, len: u32) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::random_bytes]

pub fn reseed_rng(&self, seed: Vec<u8>) -> CryptoResult<()>

see [mls_crypto_provider::MlsCryptoProvider::reseed]

pub fn commit_accepted(
&self,
conversation_id: ConversationId
) -> CryptoResult<()>

See [core_crypto::MlsCentral::commit_accepted]

pub fn clear_pending_proposal(
&self,
conversation_id: ConversationId,
proposal_ref: Vec<u8>
) -> CryptoResult<()>

See [core_crypto::MlsCentral::clear_pending_proposal]

pub fn clear_pending_commit(
&self,
conversation_id: ConversationId
) -> CryptoResult<()>

See [core_crypto::MlsCentral::clear_pending_commit]

pub fn get_client_ids(
&self,
conversation_id: ConversationId
) -> CryptoResult<Vec<ClientId>>

See [core_crypto::MlsCentral::get_client_ids]

pub fn export_secret_key(
&self,
conversation_id: ConversationId,
key_length: u32
) -> CryptoResult<Vec<u8>>

See [core_crypto::MlsCentral::export_secret_key]

impl CoreCrypto<'_>

pub fn proteus_init(&self) -> CryptoResult<()>

See [core_crypto::proteus::ProteusCentral::try_new]

pub fn proteus_session_from_prekey(
&self,
session_id: &str,
prekey: &[u8]
) -> CryptoResult<()>

See [core_crypto::proteus::ProteusCentral::session_from_prekey]

pub fn proteus_session_from_message(
&self,
session_id: &str,
envelope: &[u8]
) -> CryptoResult<Vec<u8>>

See [core_crypto::proteus::ProteusCentral::session_from_message]

pub fn proteus_session_save(&self, session_id: &str) -> CryptoResult<()>

See [core_crypto::proteus::ProteusCentral::session_save] Note: This isn’t usually needed as persisting sessions happens automatically when decrypting/encrypting messages and initializing Sessions

pub fn proteus_session_delete(&self, session_id: &str) -> CryptoResult<()>

See [core_crypto::proteus::ProteusCentral::session_delete]

pub fn proteus_session_exists(&self, session_id: &str) -> CryptoResult<bool>

See [core_crypto::proteus::ProteusCentral::session_exists]

pub fn proteus_decrypt(
&self,
session_id: &str,
ciphertext: &[u8]
) -> CryptoResult<Vec<u8>>

See [core_crypto::proteus::ProteusCentral::decrypt]

pub fn proteus_encrypt(
&self,
session_id: &str,
plaintext: &[u8]
) -> CryptoResult<Vec<u8>>

See [core_crypto::proteus::ProteusCentral::encrypt]

pub fn proteus_encrypt_batched(
&self,
sessions: Vec<String>,
plaintext: &[u8]
) -> CryptoResult<HashMap<String, Vec<u8>>>

See [core_crypto::proteus::ProteusCentral::encrypt_batched]

pub fn proteus_new_prekey(&self, prekey_id: u16) -> CryptoResult<Vec<u8>>

See [core_crypto::proteus::ProteusCentral::new_prekey]

pub fn proteus_new_prekey_auto(&self) -> CryptoResult<Vec<u8>>

See [core_crypto::proteus::ProteusCentral::new_prekey_auto]

pub fn proteus_fingerprint(&self) -> CryptoResult<String>

See [core_crypto::proteus::ProteusCentral::fingerprint]

pub fn proteus_fingerprint_local(&self, session_id: &str) -> CryptoResult<String>

See [core_crypto::proteus::ProteusCentral::fingerprint_local]

pub fn proteus_fingerprint_remote(
&self,
session_id: &str
) -> CryptoResult<String>

See [core_crypto::proteus::ProteusCentral::fingerprint_remote]

pub fn proteus_fingerprint_prekeybundle(
&self,
prekey: &[u8]
) -> CryptoResult<String>

See [core_crypto::proteus::ProteusCentral::fingerprint_prekeybundle] NOTE: uniffi doesn’t support associated functions, so we have to have the self here

pub fn proteus_cryptobox_migrate(&self, path: &str) -> CryptoResult<()>

See [core_crypto::proteus::ProteusCentral::cryptobox_migrate]