core_crypto_ffi

Struct CoreCryptoContext

pub struct CoreCryptoContext { /* private fields */ }

The CoreCryptoContext holds the primary CoreCrypto APIs.

An instance of this struct is provided to the function handed to CoreCrypto::transaction.

Every mutable operation is done through this struct. This struct will buffer all operations in memory and when [TransactionContext::finish] is called, it will persist the data into the keystore.

Implementations

impl CoreCryptoContext

pub async fn e2ei_new_enrollment( &self, client_id: String, display_name: String, handle: String, expiry_sec: u32, ciphersuite: Ciphersuite, team: Option<String>, ) -> CoreCryptoResult<E2eiEnrollment>

See [core_crypto::transaction_context::TransactionContext::e2ei_new_enrollment]

pub async fn e2ei_register_acme_ca( &self, trust_anchor_pem: String, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::e2ei_register_acme_ca]

pub async fn e2ei_register_intermediate_ca( &self, cert_pem: String, ) -> CoreCryptoResult<Option<Vec<String>>>

See [core_crypto::transaction_context::TransactionContext::e2ei_register_intermediate_ca_pem]

pub async fn e2ei_register_crl( &self, crl_dp: String, crl_der: Vec<u8>, ) -> CoreCryptoResult<CrlRegistration>

See [core_crypto::transaction_context::TransactionContext::e2ei_register_crl]

pub async fn e2ei_mls_init_only( &self, enrollment: Arc<E2eiEnrollment>, certificate_chain: String, transport: Arc<dyn MlsTransport>, ) -> CoreCryptoResult<CredentialRef>

See [core_crypto::transaction_context::TransactionContext::e2ei_mls_init_only]

pub async fn save_x509_credential( &self, enrollment: Arc<E2eiEnrollment>, certificate_chain: String, ) -> CoreCryptoResult<CredentialRef>

See [core_crypto::transaction_context::TransactionContext::save_x509_credential]

pub async fn e2ei_conversation_state( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<E2eiConversationState>

See [core_crypto::mls::conversation::Conversation::e2ei_conversation_state]

pub async fn e2ei_is_enabled( &self, ciphersuite: Ciphersuite, ) -> CoreCryptoResult<bool>

See [core_crypto::Session::e2ei_is_enabled]

pub async fn get_device_identities( &self, conversation_id: &ConversationId, device_ids: Vec<Arc<ClientId>>, ) -> CoreCryptoResult<Vec<WireIdentity>>

See [core_crypto::mls::conversation::Conversation::get_device_identities]

pub async fn get_user_identities( &self, conversation_id: &ConversationId, user_ids: Vec<String>, ) -> CoreCryptoResult<HashMap<String, Vec<WireIdentity>>>

See [core_crypto::mls::conversation::Conversation::get_user_identities]

pub async fn e2ei_is_pki_env_setup(&self) -> bool

See [core_crypto::Session::e2ei_is_pki_env_setup]

impl CoreCryptoContext

pub async fn mls_init( &self, client_id: &Arc<ClientId>, transport: Arc<dyn MlsTransport>, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::mls_init]

pub async fn conversation_epoch( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<u64>

See [core_crypto::mls::conversation::Conversation::epoch]

pub async fn conversation_ciphersuite( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<Ciphersuite>

See [core_crypto::mls::conversation::Conversation::ciphersuite]

pub async fn conversation_credential( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<CredentialRef>

Get the credential ref for the given conversation.

pub async fn set_conversation_credential( &self, conversation_id: &ConversationId, credential_ref: Arc<CredentialRef>, ) -> CoreCryptoResult<()>

Set the credential ref for the given conversation.

pub async fn conversation_exists( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<bool>

See [core_crypto::Session::conversation_exists]

pub async fn get_client_ids( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<Vec<Arc<ClientId>>>

See [core_crypto::mls::conversation::Conversation::get_client_ids]

pub async fn export_secret_key( &self, conversation_id: &ConversationId, key_length: u32, ) -> CoreCryptoResult<SecretKey>

See [core_crypto::mls::conversation::Conversation::export_secret_key]

pub async fn get_external_sender( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<ExternalSenderKey>

See [core_crypto::mls::conversation::Conversation::get_external_sender]

pub async fn create_conversation( &self, conversation_id: &ConversationId, credential_ref: &CredentialRef, external_sender: Option<Arc<ExternalSenderKey>>, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::new_conversation]

pub async fn process_welcome_message( &self, welcome_message: Arc<Welcome>, ) -> CoreCryptoResult<WelcomeBundle>

See [core_crypto::transaction_context::TransactionContext::process_raw_welcome_message]

pub async fn add_clients_to_conversation( &self, conversation_id: &ConversationId, key_packages: Vec<Arc<Keypackage>>, ) -> CoreCryptoResult<Option<Vec<String>>>

See [core_crypto::mls::conversation::ConversationGuard::add_members]

pub async fn remove_clients_from_conversation( &self, conversation_id: &ConversationId, clients: Vec<Arc<ClientId>>, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::remove_members]

pub async fn mark_conversation_as_child_of( &self, child_id: &ConversationId, parent_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::mark_as_child_of]

pub async fn update_keying_material( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::update_key_material]

pub async fn commit_pending_proposals( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::commit_pending_proposals]

pub async fn wipe_conversation( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::wipe]

pub async fn decrypt_message( &self, conversation_id: &ConversationId, payload: Vec<u8>, ) -> CoreCryptoResult<DecryptedMessage>

See [core_crypto::mls::conversation::ConversationGuard::decrypt_message]

pub async fn encrypt_message( &self, conversation_id: &ConversationId, message: Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::mls::conversation::ConversationGuard::encrypt_message]

pub async fn join_by_external_commit( &self, group_info: Arc<GroupInfo>, credential_ref: Arc<CredentialRef>, ) -> CoreCryptoResult<WelcomeBundle>

See [core_crypto::transaction_context::TransactionContext::join_by_external_commit]

pub async fn enable_history_sharing( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::enable_history_sharing]

pub async fn disable_history_sharing( &self, conversation_id: &ConversationId, ) -> CoreCryptoResult<()>

See [core_crypto::mls::conversation::ConversationGuard::disable_history_sharing]

pub async fn add_credential( &self, credential: Arc<Credential>, ) -> CoreCryptoResult<CredentialRef>

Add a Credential to this client.

Note that while an arbitrary number of credentials can be generated, those which are added to a CC instance must be distinct in credential type, signature scheme, and the timestamp of creation. This timestamp has only 1 second of resolution, limiting the number of credentials which can be added. This is a known limitation and will be relaxed in the future.

pub async fn remove_credential( &self, credential_ref: &Arc<CredentialRef>, ) -> CoreCryptoResult<()>

Remove a Credential from this client.

pub async fn get_credentials(&self) -> CoreCryptoResult<Vec<Arc<CredentialRef>>>

Get all credentials from this client.

pub async fn find_credentials( &self, client_id: Option<Arc<ClientId>>, public_key: Option<Vec<u8>>, ciphersuite: Option<Ciphersuite>, credential_type: Option<CredentialType>, earliest_validity: Option<u64>, ) -> CoreCryptoResult<Vec<Arc<CredentialRef>>>

Get all credentials from this client which match the provided parameters.

Parameters which are unset or None match anything. Those with a particular value find only credentials matching that value.

pub async fn generate_keypackage( &self, credential_ref: &Arc<CredentialRef>, lifetime: Option<Duration>, ) -> CoreCryptoResult<Arc<Keypackage>>

Generate a KeyPackage from the referenced credential.

Makes no attempt to look up or prune existing keypackges.

If lifetime is set, the keypackages will expire that span into the future. If it is unset, a default lifetime of approximately 3 months is used.

pub async fn get_keypackages(&self) -> CoreCryptoResult<Vec<Arc<KeypackageRef>>>

Get a reference to each KeyPackage in the database.

pub async fn remove_keypackage( &self, kp_ref: &Arc<KeypackageRef>, ) -> CoreCryptoResult<()>

Remove a KeyPackage from the database.

pub async fn remove_keypackages_for( &self, credential_ref: &Arc<CredentialRef>, ) -> CoreCryptoResult<()>

Remove all KeyPackages associated with this credential ref.

impl CoreCryptoContext

pub async fn proteus_init(&self) -> CoreCryptoResult<()>

See [core_crypto::proteus::ProteusCentral::try_new]

pub async fn proteus_session_from_prekey( &self, session_id: String, prekey: Vec<u8>, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::proteus_session_from_prekey]

pub async fn proteus_session_from_message( &self, session_id: &str, envelope: Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::transaction_context::TransactionContext::proteus_session_from_message]

pub async fn proteus_session_save( &self, session_id: &str, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::proteus_session_save]

Note: This isnโt usually needed as persisting sessions happens automatically when decrypting/encrypting messages and initializing Sessions

pub async fn proteus_session_delete( &self, session_id: String, ) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::proteus_session_delete]

pub async fn proteus_session_exists( &self, session_id: &str, ) -> CoreCryptoResult<bool>

See [core_crypto::transaction_context::TransactionContext::proteus_session_exists]

pub async fn proteus_decrypt( &self, session_id: &str, ciphertext: Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::transaction_context::TransactionContext::proteus_decrypt]

pub async fn proteus_decrypt_safe( &self, session_id: &str, ciphertext: Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

Decrypt a message whether or not the proteus session already exists, and saves the session.

This is intended to replace simple usages of proteusDecrypt.

However, when decrypting large numbers of messages in a single session, the existing methods may be more efficient.

pub async fn proteus_encrypt( &self, session_id: String, plaintext: Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::transaction_context::TransactionContext::proteus_encrypt]

pub async fn proteus_encrypt_batched( &self, sessions: Vec<String>, plaintext: Vec<u8>, ) -> CoreCryptoResult<HashMap<String, Vec<u8>>>

See [core_crypto::transaction_context::TransactionContext::proteus_encrypt_batched]

pub async fn proteus_new_prekey( &self, prekey_id: u16, ) -> CoreCryptoResult<Vec<u8>>

Creates a new Proteus prekey with the given id and returns the CBOR-serialized version of the prekey bundle

Warning: The Proteus client MUST be initialized with proteus_init first or an error will be returned

See [core_crypto::transaction_context::TransactionContext::proteus_new_prekey]

pub async fn proteus_new_prekey_auto( &self, ) -> CoreCryptoResult<ProteusAutoPrekeyBundle>

Creates a new Proteus prekey with an automatically incremented ID and returns the CBOR-serialized version of the prekey bundle

Warning: The Proteus client MUST be initialized with proteus_init first or an error will be returned

See [core_crypto::transaction_context::TransactionContext::proteus_new_prekey_auto]

pub async fn proteus_last_resort_prekey(&self) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::transaction_context::TransactionContext::proteus_last_resort_prekey]

pub async fn proteus_fingerprint(&self) -> CoreCryptoResult<String>

See [core_crypto::transaction_context::TransactionContext::proteus_fingerprint]

pub async fn proteus_fingerprint_local( &self, session_id: String, ) -> CoreCryptoResult<String>

See [core_crypto::transaction_context::TransactionContext::proteus_fingerprint_local]

pub async fn proteus_fingerprint_remote( &self, session_id: String, ) -> CoreCryptoResult<String>

See [core_crypto::transaction_context::TransactionContext::proteus_fingerprint_remote]

pub async fn proteus_reload_sessions(&self) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::proteus_reload_sessions]

impl CoreCryptoContext

pub fn new(instance: Arc<Self>) -> Arc<Self>

This is only needed to allow TS inheritance and should be hidden from library consumers.

impl CoreCryptoContext

pub async fn set_data(&self, data: Vec<u8>) -> CoreCryptoResult<()>

See [core_crypto::transaction_context::TransactionContext::set_data]

pub async fn get_data(&self) -> CoreCryptoResult<Option<Vec<u8>>>

See [core_crypto::transaction_context::TransactionContext::get_data]

pub async fn random_bytes(&self, len: u32) -> CoreCryptoResult<Vec<u8>>

See [core_crypto::Session::random_bytes].