Trait core_crypto_ffi::CoreCryptoCallbacks

pub trait CoreCryptoCallbacks: Debug + Send + Sync {
    fn authorize(
        &self,
        conversation_id: Vec<u8, Global>,
        client_id: ClientId
    ) -> bool;
    fn user_authorize(
        &self,
        conversation_id: Vec<u8, Global>,
        external_client_id: ClientId,
        existing_clients: Vec<ClientId, Global>
    ) -> bool;
    fn client_is_existing_group_user(
        &self,
        client_id: ClientId,
        existing_clients: Vec<ClientId, Global>
    ) -> bool;
}

This trait is used to provide callback mechanisms for the MlsCentral struct, for example for operations like adding or removing memebers that can be authorized through a caller provided authorization method.

Required Methods

fn authorize(&self, conversation_id: Vec<u8, Global>, client_id: ClientId) -> bool

Function responsible for authorizing an operation. Returns true if the operation is authorized.

Arguments

conversation_id - id of the group/conversation

client_id - id of the client to authorize

fn user_authorize(
&self,
conversation_id: Vec<u8, Global>,
external_client_id: ClientId,
existing_clients: Vec<ClientId, Global>
) -> bool

Function responsible for authorizing an operation for a given user. Use [external_client_id] & [existing_clients] to get all the ‘client_id’ belonging to the same user as [external_client_id]. Then, given those client ids, verify that at least one has the right role (is authorized) exactly like it’s done in [authorize] Returns true if the operation is authorized.

Arguments

conversation_id - id of the group/conversation

external_client_id - id a client external to the MLS group

existing_clients - all the clients in the MLS group

fn client_is_existing_group_user(
&self,
client_id: ClientId,
existing_clients: Vec<ClientId, Global>
) -> bool

Validates if the given client_id belongs to one of the provided existing_clients This basically allows to defer the client ID parsing logic to the caller - because CoreCrypto is oblivious to such things

Arguments

client_id - client ID of the client referenced within the sent proposal

existing_clients - all the clients in the MLS group