Trait core_crypto_ffi::Serialize

pub trait Serialize: Size {
    fn tls_serialize<W>(&self, writer: &mut W) -> Result<usize, Error>
    where
        W: Write;

    fn tls_serialize_detached(&self) -> Result<Vec<u8, Global>, Error> { ... }
}

The Serialize trait provides functions to serialize a struct or enum.

The trait provides two functions:

tls_serialize that takes a buffer to write the serialization to

tls_serialize_detached that returns a byte vector

Required Methods

fn tls_serialize<W>(&self, writer: &mut W) -> Result<usize, Error>where
W: Write,

Serialize self and write it to the writer. The function returns the number of bytes written to writer.

Provided Methods

fn tls_serialize_detached(&self) -> Result<Vec<u8, Global>, Error>

Serialize self and return it as a byte vector.