Crate core_crypto_ffi

FFI bindings for core-crypto.

Actual implementation happens in the core-crypto crate. This crate is about setting up all the necessary annotations, wrappers, etc necessary to package those types, items, and methods for FFI via uniffi and wasm-bindgen.

Macros

proteus_impl

Implement a proteus function, or return an error if the feature is not enabled.

Structs

AcmeChallenge

For creating a challenge.

AcmeDirectory

Holds URLs of all the standard ACME endpoint supported on an ACME server.

BufferedDecryptedMessage

to avoid recursive structs

BuildMetadata

Metadata describing the conditions of the build of this software.

ClientId

A Client identifier

CommitBundle

Information returned when a commit is created.

ConversationId

A unique identifier for a single conversation.

CoreCryptoContext

The CoreCryptoContext holds the primary CoreCrypto APIs.

CoreCryptoFfi

CoreCrypto wraps around MLS and Proteus implementations and provides a transactional interface for each.

Credential

A cryptographic credential.

CredentialRef

A reference to a credential which has been persisted in CC.

CrlRegistration

Supporting struct for CRL registration result

Database

The database acting as a core crypto keystore.

DatabaseKey

The key used to encrypt the database.

DecryptedMessage

A decrypted message and various associated metadata

E2eiEnrollment

Wire end to end identity solution for fetching a x509 certificate which identifies a client.

GroupInfoBundle

A GroupInfo with some metadata

HistorySecret

A HistorySecret encodes sufficient client state that it can be used to instantiate an ephemeral client.

HttpHeader

An HttpHeader used for pki hooks

HttpResponse

An HttpResponse used for pki hooks

Keypackage

A signed object describing a client’s identity and capabilities.

KeypackageRef

A lightweight distinct reference to a KeyPackage sufficient to uniquely identify it

MlsTransportData

An entity / data which has been packaged by the application to be encrypted and transmitted in an application message.

NewAcmeAuthz

Result of an authorization creation.

NewAcmeOrder

Result of an order creation.

PkiEnvironment

A PkiEnvironment

ProteusAutoPrekeyBundle

Encapsulates a prekey id and a cbor-serialized prekey

Timestamp

A wrapper around SystemTime for FFI bindings with custom type mapping per language.

TransactionHelper

Helper for working with the new transasction interface.

WelcomeBundle

see [core_crypto::WelcomeBundle]

WireIdentity

Represents the identity claims identifying a client Those claims are verifiable by any member in the group

X509Identity

Represents the parts of WireIdentity that are specific to a X509 certificate (and not a Basic one).

Enums

Ciphersuite

MLS ciphersuites.

CoreCryptoLogLevel

Defines the log level for a CoreCrypto

CredentialType

Type of Credential

DeviceStatus

Indicates the standalone status of a device Credential in a MLS group at a moment T.

E2eiConversationState

Indicates the state of a Conversation regarding end-to-end identity.

HttpMethod

HttpMethod used for pki hooks

MlsError

MLS produces these kinds of error

MlsTransportResponse

MLS transport may or may not succeeed; this response indicates to CC the outcome of the transport attempt.

ProteusError

Proteus produces these kinds of error

SignatureScheme

MLS signature schemes

Traits

CoreCryptoCommand

A CoreCryptoCommand has an execute method which accepts a CoreCryptoContext and returns nothing.

CoreCryptoLogger

This trait is used to provide a callback mechanism to hook up the respective platform logging system.

EpochObserver

An EpochObserver is notified whenever a conversation’s epoch changes.

MlsTransport

Used by core crypto to send commits or application messages to the delivery service. This trait must be implemented before calling any functions that produce commits.

PkiEnvironmentHooks

The PKI Environment Hooks used for external calls during e2e enrollment flow. When communicating with the Identity Provider (IDP) and Wire server, CoreCrypto delegates to the client app by calling the relevant methods.

Functions

build_metadata

Returns build data for CoreCrypto

ciphersuite_default

Get an instance of the default ciphersuite.

ciphersuite_from_u16

Construct a ciphersuite enum instance from its discriminant.

core_crypto_history_client

Instantiate a history client.

create_pki_environment

Create a new PKI environment

credential_basic

Generate a basic credential.

in_memory_database

Create an in-memory Database whose data will be lost when the instance is dropped.

migrate_database_key_type_to_bytes

Updates the key of the CoreCrypto database. To be used only once, when moving from CoreCrypto <= 5.x to CoreCrypto 6.x.

open_database

Open or create a Database.

set_logger

Initializes the logger

set_max_log_level

Set maximum log level forwarded to the logger

update_database_key

Updates the key of the CoreCrypto database.

version

The version of core-crypto.

Type Aliases

CoreCryptoError

The primary error type for CoreCrypto

CoreCryptoResult

A result which produces a CoreCryptoError by default.