Crate core_crypto_ffi

Modules

bindings

CoreCrypto FFI Details

context

Macros

proteus_impl

Structs

AcmeChallenge

See core_crypto::e2e_identity::types::E2eiAcmeChallenge

AcmeDirectory

See core_crypto::e2e_identity::types::E2eiAcmeDirectory

BufferedDecryptedMessage

because Uniffi does not support recursive structs

BuildMetadata

Metadata describing the conditions of the build of this software.

Ciphersuite

Ciphersuites

ClientId

CommitBundle

ConversationConfiguration

See core_crypto::prelude::MlsConversationConfiguration

CoreCrypto

CrlRegistration

Supporting struct for CRL registration result

CustomConfiguration

See core_crypto::prelude::MlsCustomConfiguration

DecryptedMessage

See [core_crypto::prelude::decrypt::MlsConversationDecryptMessage]

E2eiDumpedPkiEnv

Dummy comment

E2eiEnrollment

See core_crypto::e2e_identity::E2eiEnrollment

GroupInfoBundle

NewAcmeAuthz

See core_crypto::e2e_identity::types::E2eiNewAcmeAuthz

NewAcmeOrder

See core_crypto::e2e_identity::types::E2eiNewAcmeOrder

NewCrlDistributionPoints

ProposalBundle

ProteusAutoPrekeyBundle

UniFfiTraitVtableCoreCryptoLogger

UniFfiTraitVtableMlsTransport

WelcomeBundle

see core_crypto::prelude::MlsConversationCreationMessage

WireIdentity

See core_crypto::prelude::WireIdentity

X509Identity

See core_crypto::prelude::X509Identity

Enums

CiphersuiteName

CoreCryptoError

CoreCryptoLogLevel

Defines the log level for a CoreCrypto

DeviceStatus

E2eiConversationState

MlsCredentialType

MlsError

MlsGroupInfoEncryptionType

MlsRatchetTreeType

MlsTransportResponse

MlsWirePolicy

ProteusError

Traits

CoreCryptoLogger

This trait is used to provide a callback mechanism to hook up the rerspective platform logging system

MlsTransport

Used by core crypto to send commits or application messages to the delivery service. This trait must be implemented before calling any functions that produce commits.

Functions

build_metadata

core_crypto_deferred_init

Similar to core_crypto_new but defers MLS initialization. It can be initialized later with CoreCryptoContext::mls_init.

core_crypto_new

See [core_crypto::mls::Client::try_new]

set_logger

Initializes the logger

set_logger_only

Initializes the logger

set_max_log_level

Set maximum log level forwarded to the logger

uniffi_core_crypto_ffi_fn_init_callback_vtable_corecryptologger

uniffi_core_crypto_ffi_fn_init_callback_vtable_mlstransport

version

Type Aliases

ConversationId

A unique identifier for a group/conversation. The identifier must be unique within a client.