Crate core_crypto_ffi

Modules

bindings

CoreCrypto FFI Details

Macros

conversation_id_vec

proteus_impl

Implement a proteus function, or return an error if the feature is not enabled.

Structs

AcmeChallenge

For creating a challenge.

AcmeDirectory

Holds URLs of all the standard ACME endpoint supported on an ACME server.

BufferedDecryptedMessage

to avoid recursive structs

BuildMetadata

Metadata describing the conditions of the build of this software.

Ciphersuite

Ciphersuites

ClientId

CommitBundle

ConversationConfiguration

See core_crypto::prelude::MlsConversationConfiguration

CoreCrypto

CoreCryptoContext

CrlRegistration

Supporting struct for CRL registration result

CustomConfiguration

see core_crypto::prelude::MlsCustomConfiguration

DatabaseKey

DecryptedMessage

See [core_crypto::prelude::decrypt::MlsConversationDecryptMessage]

E2eiDumpedPkiEnv

Dump of the PKI environemnt as PEM

E2eiEnrollment

See core_crypto::e2e_identity::E2eiEnrollment

GroupInfoBundle

NewAcmeAuthz

Result of an authorization creation.

NewAcmeOrder

Result of an order creation.

NewCrlDistributionPoints

ProposalBundle

ProteusAutoPrekeyBundle

TransactionHelper

Helper for working with the new transasction interface.

WelcomeBundle

see [core_crypto::prelude::MlsConversationCreationMessage]

WireIdentity

See core_crypto::prelude::WireIdentity

X509Identity

See core_crypto::prelude::X509Identity

Enums

CoreCryptoLogLevel

Defines the log level for a CoreCrypto

CredentialType

DeviceStatus

E2eiConversationState

MlsError

MlsTransportResponse

ProteusError

WirePolicy

See core_crypto::prelude::MlsWirePolicy

Traits

CoreCryptoCommand

CoreCryptoLogger

This trait is used to provide a callback mechanism to hook up the respective platform logging system.

EpochObserver

An EpochObserver is notified whenever a conversation’s epoch changes.

MlsTransport

Used by core crypto to send commits or application messages to the delivery service. This trait must be implemented before calling any functions that produce commits.

Functions

build_metadata

Returns build data for CoreCrypto

core_crypto_deferred_init

Free function to construct a new CoreCrypto instance.

core_crypto_new

Free function to construct a new CoreCrypto instance.

migrate_db_key_type_to_bytes

Updates the key of the CoreCrypto database. To be used only once, when moving from CoreCrypto <= 5.x to CoreCrypto 6.x.

set_logger

Initializes the logger

set_logger_only

Initializes the logger

set_max_log_level

Set maximum log level forwarded to the logger

version

Type Aliases

CoreCryptoError

CoreCryptoResult

FfiClientId