core_crypto_ffi

Struct CoreCrypto

pub struct CoreCrypto { /* private fields */ }

Implementations

impl CoreCrypto

pub async fn client_public_key( &self, ciphersuite: Ciphersuite, credential_type: CredentialType, ) -> CoreCryptoResult<Vec<u8>>

See core_crypto::transaction_context::TransactionContext::client_public_key

impl CoreCrypto

pub async fn transaction( &self, command: Arc<dyn CoreCryptoCommand>, ) -> CoreCryptoResult<()>

Starts a new transaction in Core Crypto. If the callback succeeds, it will be committed, otherwise, every operation performed with the context will be discarded.

When calling this function from within Rust, async functions accepting a context implement CoreCryptoCommand, so operations can be defined inline as follows:

core_crypto.transaction(Arc::new(async |context| {
    // your implementation here
    Ok(())
}))?;

impl CoreCrypto

pub async fn conversation_epoch( &self, conversation_id: &Vec<u8>, ) -> CoreCryptoResult<u64>

See core_crypto::mls::conversation::Conversation::epoch

pub async fn conversation_ciphersuite( &self, conversation_id: &Vec<u8>, ) -> CoreCryptoResult<Ciphersuite>

See core_crypto::mls::conversation::Conversation::ciphersuite

pub async fn conversation_exists( &self, conversation_id: &Vec<u8>, ) -> CoreCryptoResult<bool>

See core_crypto::prelude::Session::conversation_exists

pub async fn get_client_ids( &self, conversation_id: &Vec<u8>, ) -> CoreCryptoResult<Vec<ClientId>>

See core_crypto::mls::conversation::Conversation::get_client_ids

pub async fn get_external_sender( &self, conversation_id: &Vec<u8>, ) -> CoreCryptoResult<Vec<u8>>

See core_crypto::mls::conversation::Conversation::get_external_sender

pub async fn export_secret_key( &self, conversation_id: &Vec<u8>, key_length: u32, ) -> CoreCryptoResult<Vec<u8>>

See core_crypto::mls::conversation::Conversation::export_secret_key

impl CoreCrypto

pub async fn get_device_identities( &self, conversation_id: &Vec<u8>, device_ids: Vec<ClientId>, ) -> CoreCryptoResult<Vec<WireIdentity>>

See core_crypto::mls::conversation::Conversation::get_device_identities

pub async fn get_user_identities( &self, conversation_id: &Vec<u8>, user_ids: Vec<String>, ) -> CoreCryptoResult<HashMap<String, Vec<WireIdentity>>>

See core_crypto::mls::conversation::Conversation::get_user_identities

pub async fn get_credential_in_use( &self, group_info: Vec<u8>, credential_type: CredentialType, ) -> CoreCryptoResult<E2eiConversationState>

See core_crypto::prelude::Session::get_credential_in_use

impl CoreCrypto

pub async fn e2ei_dump_pki_env( &self, ) -> CoreCryptoResult<Option<E2eiDumpedPkiEnv>>

pub async fn e2ei_is_pki_env_setup(&self) -> bool

See core_crypto::prelude::Session::e2ei_is_pki_env_setup

pub async fn e2ei_is_enabled( &self, ciphersuite: Ciphersuite, ) -> CoreCryptoResult<bool>

See core_crypto::prelude::Session::e2ei_is_enabled

impl CoreCrypto

pub async fn register_epoch_observer( &self, epoch_observer: Arc<dyn EpochObserver>, ) -> CoreCryptoResult<()>

Add an epoch observer to this client.

This function should be called 0 or 1 times in a client’s lifetime. If called when an epoch observer already exists, this will return an error.

impl CoreCrypto

pub async fn provide_transport( &self, callbacks: Arc<dyn MlsTransport>, ) -> CoreCryptoResult<()>

See core_crypto::prelude::Session::provide_transport

impl CoreCrypto

pub async fn proteus_session_exists( &self, session_id: String, ) -> CoreCryptoResult<bool>

See core_crypto::proteus::ProteusCentral::session_exists

pub async fn proteus_fingerprint(&self) -> CoreCryptoResult<String>

See core_crypto::proteus::ProteusCentral::fingerprint

pub async fn proteus_fingerprint_local( &self, session_id: String, ) -> CoreCryptoResult<String>

See core_crypto::proteus::ProteusCentral::fingerprint_local

pub async fn proteus_fingerprint_remote( &self, session_id: String, ) -> CoreCryptoResult<String>

See core_crypto::proteus::ProteusCentral::fingerprint_remote

impl CoreCrypto

pub fn proteus_last_resort_prekey_id(&self) -> CoreCryptoResult<u16>

See core_crypto::proteus::ProteusCentral::last_resort_prekey_id

pub fn proteus_fingerprint_prekeybundle( &self, prekey: Vec<u8>, ) -> CoreCryptoResult<String>

See core_crypto::proteus::ProteusCentral::fingerprint_prekeybundle

impl CoreCrypto

pub async fn random_bytes(&self, len: u32) -> CoreCryptoResult<Vec<u8>>

See core_crypto::prelude::Session::random_bytes

pub async fn reseed_rng(&self, seed: Vec<u8>) -> CoreCryptoResult<()>

see core_crypto::prelude::Session::reseed

impl CoreCrypto

pub async fn new( path: String, key: DatabaseKey, client_id: Option<ClientId>, ciphersuites: Option<Ciphersuites>, entropy_seed: Option<Vec<u8>>, nb_key_package: Option<u32>, ) -> CoreCryptoResult<Self>

impl CoreCrypto

pub async fn can_close(&self) -> bool

See Session::can_close